(refactor): split build tests into separate files per fixture
- better organized that way as the tests are somewhat independent of
  one another and previously even I was confused where to put different
  types of tests

- also refactor some test naming and some redundant usage of flags
  - especially for 0-config, where flags shouldn't be used

- and remove testEnvironment comment as it's already set in jest config

(optim): now that it's one fixture per file, we only need to copy the
fixture directory once per file, which should be a speed boost
- and sometimes may not need to re-run `tsdx build` when not necessary
- also the different test files get parallelized, which is also faster

- before: 53s avg on my machine, after: 37s avg on my machine
  - avg for running all tests
  - pretty BIG difference
